+++ packages/graphql-transformer-core/src/__tests__/util/amplifyUtils.test.ts
@@ -1,4 +1,5 @@
-import { getSanityCheckRules, SanityCheckRules } from '../../util/amplifyUtils';
+import { getS3KeyNamesFromDirectory, getSanityCheckRules, SanityCheckRules } from '../../util/amplifyUtils';
+import mock from 'mock-fs';
 
 const buildMockedFeatureFlags = (flagValue: boolean) => {
   return {
@@ -46,5 +47,37 @@ describe('get sanity check rules', () => {
     const projectRulesFn = sanityCheckRules.projectRules.map((func) => func.name);
     expect(diffRulesFn).toMatchSnapshot();
     expect(projectRulesFn).toMatchSnapshot();
+  });
+});
+
+describe('get S3 keys from directory', () => {
+  const MOCK_ROOT_DIR = 'rootDir';
+
+  afterEach(() => {
+    mock.restore();
+  });
+
+  it('should have correct S3 keys generated from glob function', () => {
+    const mockFilePath = {};
+    mockFilePath[MOCK_ROOT_DIR] = {
+      a: {
+        b: {
+          testFile1: 'hello world',
+        },
+      },
+      c: {
+        testFile2: 'hello world',
+      },
+      testFile3: 'hello world',
+    };
+    mock(mockFilePath);
+    const keys = getS3KeyNamesFromDirectory(MOCK_ROOT_DIR);
+    expect(keys).toMatchInlineSnapshot(`
+      Array [
+        "testFile3",
+        "c/testFile2",
+        "a/b/testFile1",
+      ]
+    `);
   });
 });