+++ packages/amplify-e2e-core/CHANGELOG.md
@@ -3,6 +3,10 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [4.4.6](https://github.com/aws-amplify/amplify-category-api/compare/amplify-category-api-e2e-core@4.4.5...amplify-category-api-e2e-core@4.4.6) (2024-01-22)
+
+**Note:** Version bump only for package amplify-category-api-e2e-core
+
 ## [4.4.5](https://github.com/aws-amplify/amplify-category-api/compare/amplify-category-api-e2e-core@4.4.4...amplify-category-api-e2e-core@4.4.5) (2023-12-18)
 
 **Note:** Version bump only for package amplify-category-api-e2e-core