+++ packages/amplify-graphql-http-transformer/CHANGELOG.md
@@ -3,6 +3,14 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [2.1.16](https://github.com/aws-amplify/amplify-category-api/compare/@aws-amplify/graphql-http-transformer@2.1.15...@aws-amplify/graphql-http-transformer@2.1.16) (2024-01-30)
+
+**Note:** Version bump only for package @aws-amplify/graphql-http-transformer
+
+## [2.1.15](https://github.com/aws-amplify/amplify-category-api/compare/@aws-amplify/graphql-http-transformer@2.1.14...@aws-amplify/graphql-http-transformer@2.1.15) (2024-01-22)
+
+**Note:** Version bump only for package @aws-amplify/graphql-http-transformer
+
 ## [2.1.14](https://github.com/aws-amplify/amplify-category-api/compare/@aws-amplify/graphql-http-transformer@2.1.13...@aws-amplify/graphql-http-transformer@2.1.14) (2023-12-18)
 
 **Note:** Version bump only for package @aws-amplify/graphql-http-transformer